Create CRUD BE (News)

+++ src/modules/news/news.service.ts
@@ -10,26 +10,32 @@ import { News } from 'src/entities/News.entity';
 @Injectable()
 export class NewsService {
   constructor(
-    private NewsRepository: NewsRepository,
+    private newsRepository: NewsRepository,
     private cloudinaryService: CloudinaryService,
   ) {}
+
 
   async findAll(query: any = {}): Promise<News[]> {
     return this.NewsRepository.findAll(query);
   }
 
-  async create(data: any, files: any): Promise<News> {
+  async create(data: any, file: any): Promise<News> {
     try {
-      const uploadImages = files.map((file: any) => {
+      if (!file) {
+        // Handle the case where file are undefined
+        throw new Error('No file were provided');
+      }
+
+      const uploadImage = file.map((file: any) => {
         return this.cloudinaryService.uploadImage(file);
       });
-      const uploadResult = await Promise.all(uploadImages);
+      const uploadResult = await Promise.all(uploadImage);
 
       const newNews: News = {
         ...data,
         image: uploadResult[0]?.secure_url,
       };
-      const createNews = await this.NewsRepository.create(newNews);
+      const createNews = await this.newsRepository.create(newNews);
 
       return createNews;
     } catch (error) {
@@ -38,43 +44,52 @@ export class NewsService {
   }
 
   async findById(id: string): Promise<News> {
-    const News = await this.NewsRepository.findById(id);
-    if (!News) {
+    const news = await this.newsRepository.findById(id);
+    if (!news) {
       throw new NotFoundException(`News with ID ${id} not found.`);
     }
-    return News;
+    return news;
   }
 
-  async update(id: string, data: any, files: any): Promise<News> {
+  async update(id: string, data: any, file: any): Promise<News> {
     try {
-      // Truy vấn News cũ để lấy đường dẫn ảnh cũ từ db
-      const oldNews = await this.NewsRepository.findById(id);
-      if (!oldNews) {
-        throw new Error('Image news not found');
+      if (file) {
+        // Trong trường hợp người dùng muốn đổi ảnh mới khi cập nhật
+        // Truy vấn News cũ để lấy đường dẫn ảnh cũ từ db
+        const oldNews = await this.newsRepository.findById(id);
+        if (!oldNews) {
+          throw new Error('Image news not found');
+        }
+        // Xoá ảnh cũ trên Cloudinary
+        if (oldNews.image) {
+          await this.cloudinaryService.deleteImage(oldNews.image);
+        }
+
+        const uploadImage = file.map((file: any) => {
+          return this.cloudinaryService.uploadImage(file);
+        });
+        const uploadResult = await Promise.all(uploadImage);
+
+        const newData: News = {
+          ...data,
+          image: uploadResult[0]?.secure_url,
+        };
+
+        return await this.newsRepository.update(id, newData);
+      } else {
+        // Trường hợp người dùng chỉ đổi dữ liệu không đổi ảnh mới.
+        const newData: News = {
+          ...data,
+        };
+        return await this.newsRepository.update(id, newData);
       }
-      // Xoá ảnh cũ trên Cloudinary
-      if (oldNews.image) {
-        await this.cloudinaryService.deleteImage(oldNews.image);
-      }
-
-      const uploadImages = files.map((file: any) => {
-        return this.cloudinaryService.uploadImage(file);
-      });
-      const uploadResult = await Promise.all(uploadImages);
-
-      const newData: News = {
-        ...data,
-        image: uploadResult[0]?.secure_url,
-      };
-
-      return await this.NewsRepository.update(id, newData);
     } catch (error) {
       throw error;
     }
   }
 
   async remove(id: string): Promise<void> {
-    const oldNews = await this.NewsRepository.findById(id);
+    const oldNews = await this.newsRepository.findById(id);
     if (!oldNews) {
       throw new Error('Image news not found');
     }
@@ -83,6 +98,6 @@ export class NewsService {
     if (oldNews.image) {
       await this.cloudinaryService.deleteImage(oldNews.image);
     }
-    await this.NewsRepository.remove(id);
+    await this.newsRepository.remove(id);
   }
 }
